moving heterograph index to another file
